Merge develop into main

+++ database/migrations/2025_11_30_054605_create_wallets_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->morphs('owner');
-            $table->decimal('balance', 10, 2)->default(0);
+            $table->bigInteger('balance')->default(0);
             $table->timestamps();
         });
     }
